+++ app/assets/javascripts/Components/AccountMenu/AccountMenuPane.ts
@@ -1,0 +1,6 @@
+export enum AccountMenuPane {
+  GeneralMenu,
+  SignIn,
+  Register,
+  ConfirmPassword,
+}
