+++ app/assets/javascripts/Components/AccountMenu/ConfirmPassword.tsx
@@ -4,7 +4,7 @@ import { AppState } from '@/UIModels/AppState'
 import { observer } from 'mobx-react-lite'
 import { FunctionComponent } from 'preact'
 import { useCallback, useEffect, useRef, useState } from 'preact/hooks'
-import { AccountMenuPane } from './AccountMenu'
+import { AccountMenuPane } from './AccountMenuPane'
 import { Button } from '@/Components/Button/Button'
 import { Checkbox } from '@/Components/Checkbox/Checkbox'
 import { DecoratedPasswordInput } from '@/Components/Input/DecoratedPasswordInput'
