+++ app/assets/javascripts/Components/AccountMenu/CreateAccount.tsx
@@ -3,7 +3,7 @@ import { AppState } from '@/UIModels/AppState'
 import { observer } from 'mobx-react-lite'
 import { FunctionComponent } from 'preact'
 import { StateUpdater, useCallback, useEffect, useRef, useState } from 'preact/hooks'
-import { AccountMenuPane } from './AccountMenu'
+import { AccountMenuPane } from './AccountMenuPane'
 import { Button } from '@/Components/Button/Button'
 import { DecoratedInput } from '@/Components/Input/DecoratedInput'
 import { DecoratedPasswordInput } from '@/Components/Input/DecoratedPasswordInput'
