+++ app/assets/javascripts/Components/AccountMenu/GeneralAccountMenu.tsx
@@ -6,7 +6,7 @@ import { formatLastSyncDate } from '@/Components/Preferences/Panes/Account/Sync'
 import { SyncQueueStrategy } from '@standardnotes/snjs'
 import { STRING_GENERIC_SYNC_ERROR } from '@/Strings'
 import { useCallback, useMemo, useState } from 'preact/hooks'
-import { AccountMenuPane } from './AccountMenu'
+import { AccountMenuPane } from './AccountMenuPane'
 import { FunctionComponent } from 'preact'
 import { Menu } from '@/Components/Menu/Menu'
 import { MenuItem, MenuItemSeparator, MenuItemType } from '@/Components/Menu/MenuItem'
