+++ app/assets/javascripts/Components/Footer/Footer.tsx
@@ -11,12 +11,13 @@ import {
   STRING_UPGRADE_ACCOUNT_CONFIRM_BUTTON,
 } from '@/Strings'
 import { alertDialog, confirmDialog } from '@/Services/AlertService'
-import { AccountMenu, AccountMenuPane } from '@/Components/AccountMenu/AccountMenu'
+import { AccountMenu } from '@/Components/AccountMenu/AccountMenu'
 import { AppStateEvent, EventSource } from '@/UIModels/AppState'
 import { Icon } from '@/Components/Icon/Icon'
 import { QuickSettingsMenu } from '@/Components/QuickSettingsMenu/QuickSettingsMenu'
 import { SyncResolutionMenu } from '@/Components/SyncResolutionMenu/SyncResolutionMenu'
 import { Fragment } from 'preact'
+import { AccountMenuPane } from '../AccountMenu/AccountMenuPane'
 
 type Props = {
   application: WebApplication
