+++ app/assets/javascripts/Components/Preferences/Panes/Account/Authentication.tsx
@@ -1,4 +1,3 @@
-import { AccountMenuPane } from '@/Components/AccountMenu/AccountMenu'
 import { Button } from '@/Components/Button/Button'
 import { PreferencesGroup, PreferencesSegment, Text, Title } from '@/Components/Preferences/PreferencesComponents'
 import { WebApplication } from '@/UIModels/Application'
@@ -6,6 +5,7 @@ import { AppState } from '@/UIModels/AppState'
 import { observer } from 'mobx-react-lite'
 import { FunctionComponent } from 'preact'
 import { AccountIllustration } from '@standardnotes/icons'
+import { AccountMenuPane } from '@/Components/AccountMenu/AccountMenuPane'
 
 export const Authentication: FunctionComponent<{
   application: WebApplication
