+++ app/assets/javascripts/UIModels/AppState/AppStateEvent.ts
@@ -1,0 +1,10 @@
+export enum AppStateEvent {
+  TagChanged,
+  ActiveEditorChanged,
+  PanelResized,
+  EditorFocused,
+  BeganBackupDownload,
+  EndedBackupDownload,
+  WindowDidFocus,
+  WindowDidBlur,
+}
