+++ app/assets/javascripts/UIModels/AppState/PurchaseFlowPane.ts
@@ -1,0 +1,4 @@
+export enum PurchaseFlowPane {
+  SignIn,
+  CreateAccount,
+}
