+++ app/assets/javascripts/UIModels/AppState/index.ts
@@ -1,1 +1,3 @@
-export { AppState, AppStateEvent, EventSource, PanelResizedData } from './AppState'
+export { AppState, EventSource, PanelResizedData } from './AppState'
+export * from './AppStateEvent'
+export * from './PurchaseFlowPane'
